requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0
transformers>=4.30.0
diffusers>=0.20.0
accelerate>=0.20.0
timm>=0.9.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
wandb>=0.15.0
tqdm>=4.65.0
pandas>=2.0.0
h5py>=3.9.0
seaborn>=0.12.0
scipy>=1.10.0
lpips>=0.1.4
scikit-image>=0.20.0